requirements.txt
ultralytics==8.2.0
fastapi
uvicorn[standard]
opencv-python
torch
torchvision
pillow
